Renamed encryption to encrypt.
* Added new hmac signature based functions.
  * Added Sign function, which will return a signature based on the payload and shared secret.
  * Added Validate function, which will validate the incoming message with its signature.
