+++ encrypt/keys_test.go
@@ -1,4 +1,4 @@
-package encryption
+package encrypt
 
 import (
 	"reflect"
